Code finally decomposed
